+++ actionview/test/activerecord/partial_rendering_query_test.rb
@@ -10,13 +10,10 @@ class PartialRenderingQueryTest < ActiveRecordTestCase
   end
 
   def test_render_with_relation_collection
-    notifications = capture_notifications("sql.active_record") do
-      @view.render partial: "topics/topic", collection: Topic.all
+    assert_notifications_count("sql.active_record", 1) do
+      assert_notification("sql.active_record", sql: 'SELECT "topics".* FROM "topics"') do
+        @view.render partial: "topics/topic", collection: Topic.all
+      end
     end
-
-    queries = notifications.filter_map { _1.payload[:sql] unless %w[ SCHEMA TRANSACTION ].include?(_1.payload[:name]) }
-
-    assert_equal 1, queries.size
-    assert_equal 'SELECT "topics".* FROM "topics"', queries[0]
   end
 end