+++ activestorage/test/analyzer/image_analyzer/vips_test.rb
@@ -48,13 +48,13 @@ class ActiveStorage::Analyzer::ImageAnalyzer::VipsTest < ActiveSupport::TestCase
 
   test "instrumenting analysis" do
     analyze_with_vips do
-      events = capture_notifications("analyze.active_storage") do
-        blob = create_file_blob(filename: "racecar.jpg", content_type: "image/jpeg")
-        blob.analyze
-      end
+      blob = create_file_blob(filename: "racecar.jpg", content_type: "image/jpeg")
 
-      assert_equal 1, events.size
-      assert_equal({ analyzer: "vips" }, events.first.payload)
+      assert_notifications_count("analyze.active_storage", 1) do
+        assert_notification("analyze.active_storage", analyzer: "vips") do
+          blob.analyze
+        end
+      end
     end
   end